+++ common-libs/did-auth-lib/CHANGELOG.md
@@ -1,3 +1,7 @@
+# release 2.5.0 (2022-11-14)
+* updated `url-resolver` with internal URLs for new dev environment
+# release 2.4.0 (2022-11-10)
+* updated `url-resolver` to move to new dev environment
 # release 2.3.0 (2022-09-28)
 * updated `@affinidi/common`, `internal-api-clients` and `url-resolver`
 # release 2.2.0 (2022-09-06)